Use size_t everywhere

+++ src/vector.c
@@ -19,7 +19,7 @@ Vector* CreateVector() {
   return vec;
 }
 
-void* PeekVector(Vector *vector, int index) {
+void* PeekVector(Vector *vector, size_t index) {
   if (index < 0 || index >= vector->size) {
     return NULL;
   }
